feat: Add ns-navbar component
뉴스 스탠드를 조작하는 navbar component
