+++ src/components/main/main__left/ns__container/NsContainerComponent.ts
@@ -2,6 +2,7 @@ import { Props, State } from '@src/types/types';
 import { Component } from '@src/types/interfaces';
 import { NsContainerModel } from '@components/main/main__left/ns__container/NsContainerModel.js';
 import { NsContainerView } from '@components/main/main__left/ns__container/NsContainerView.js';
+import { NsNavbarComponent } from '@components/main/main__left/ns__container/ns__navbar/NsNavbarComponent.js';
 
 export class NsContainerComponent implements Component {
   private _model: NsContainerModel;
@@ -9,6 +10,9 @@ export class NsContainerComponent implements Component {
   constructor(props?: Props) {
     this._model = new NsContainerModel();
     this._view = new NsContainerView();
+
+    const nsNavbar = new NsNavbarComponent();
+    nsNavbar.attachTo(this);
   }
 
   get element() {
