+++ src/components/main/main__left/ns__container/ns__navbar/NsNavbarModel.ts
@@ -1,6 +1,6 @@
 import { AbstractModel } from '@src/types/abstracts.js';
 
-export class NsControllerModel extends AbstractModel {
+export class NsNavbarModel extends AbstractModel {
   constructor() {
     super();
   }
